AIRelevanceFilterService/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
requests
google-generativeai
python-dotenv==1.0.1
